pyproject.toml
[build-system]
requires = ["uv_build>=0.10.8,<0.11"]
build-backend = "uv_build"

[project]
name = "auth0-llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "auth0-ai-llamaindex>=1.0.1",
    "fastapi>=0.135.1",
    "llama-cloud>=1.6.0",
    "llama-index-core>=0.14.17",
    "llama-index-indices-managed-llama-cloud>=0.11.0",
    "llama-index-llms-anthropic>=0.11.0",
    "openfga-sdk>=0.9.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]

[tool.uv.build-backend]
module-name = "auth0_llamaindex"

[project.scripts]
run-server = "auth0_llamaindex.api:run"
run-client = "auth0_llamaindex.client:main"
